+++ php/config-spec.php
@@ -19,34 +19,22 @@ return array(
     'default' => true,
     'desc' => 'Use color in output',
   ),
-  'silent' => array(
-    'runtime' => '',
-    'file' => '<bool>',
-    'default' => false,
-    'desc' => 'Suppress output',
-  ),
-  'json' => array(
-    'runtime' => '',
-    'file'  => '<bool>',
-    'default' => false,
-    'desc' => 'Force json output',
-  ),
   'debug' => array(
     'runtime' => '',
     'file'  => '<bool>',
     'default' => false,
     'desc'  => 'Get debug output',
   ),
-  'bash' => array(
-    'runtime' => '',
-    'file'  => '<bool>',
-    'default' => false,
-    'desc'  => 'Bash-friendly output',
-  ),
   'yes' => array(
     'runtime' => '',
     'file'  => '<bool>',
     'default' => false,
     'desc'  => 'Answer yes to all prompts',
+  ),
+  'format' => array(
+    'runtime' => '=<json|bash|silent>',
+    'file' => '=<json|bash|silent>',
+    'default' => 'normal',
+    'desc' => 'Force json output',
   ),
 );