+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 
 ### Changed
 - Logged errors now exit with -1. (#576)
+- `--bash`, `--json`, and `--silent` have been replaced with `--format=<bash|json|silent` (#577)
 
 ##[0.8.1] - 2015-09-28
 ### Changed